Add means for sending help to a specific user
Motivation: In some situations you may want to help some user with the bot by
having the bot directly offer help to that user.

I added an optional parameter to the `help` command, which takes a nickname and
makes the bot offer help to that specific user by mentioning him instead of the
user invoking the `help` command.